refactor: 미디어 쿼리 관련 훅 리팩토링 - mediaType 대신 breakpoint로 명시
1. 훅 이름 변경
2. 변수명 변경
