+++ frontend/src/pages/WrittenReviewPage/hooks/index.ts
@@ -1,1 +1,1 @@
-export { default as useCurrentMediaType } from './useCurrentMediaType';
+export { default as useDeviceBreakpoints } from './useDeviceBreakpoints';
